computer_vision_api/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
keras==2.14.0
tensorflow==2.14.0
numpy==1.25.0
Pillow==10.3.0
